Corregido bug con el manejo de mapas

+++ ReduceFatFast/src/main/java/ar/com/reduceFatFast/controller/GrupoController.java
@@ -172,7 +172,7 @@ public class GrupoController extends AbstractController {
 	}
 	)
     @RequestMapping(path="/grupos/{idGrupo}/dieta/comidas", method = RequestMethod.POST)
-    public ResponseEntity<String> agregarComidaADieta(@PathVariable("idGrupo") long idGrupo, long idComida, Integer dia, int comidaDelDia){
+    public ResponseEntity<String> agregarComidaADieta(@PathVariable("idGrupo") long idGrupo, long idComida, Long dia, int comidaDelDia){
     	this.getGrupoService().agregarComidaADieta(idGrupo, idComida, dia, comidaDelDia);
 		
     	return new ResponseEntity<>(HttpStatus.OK);
