+++ ReduceFatFast/src/main/java/ar/com/reduceFatFast/dto/DiaDto.java
@@ -15,18 +15,18 @@ import lombok.Data;
  */
 public @Data class DiaDto {
 	public long id;
-	public int numero;
+	public Long numero;
 	public Integer cantidadComidasPorDia;
 	public List<ComidaDto> comidas;
 	
-	public DiaDto(Dia dia, int numero) {
+	public DiaDto(Dia dia, Long numero) {
 		super();
 		this.setId(dia.getId());
 		this.setNumero(numero);
 		this.setCantidadComidasPorDia(dia.getCantidadComidasPorDia());
 		this.setComidas(new ArrayList<ComidaDto>());
 		
-		for(Integer each : dia.getComidas().keySet()) {
+		for(Long each : dia.getComidas().keySet()) {
 			this.getComidas().add(new ComidaDto(dia.getComidas().get(each)));
 		}
 	}
