+++ ReduceFatFast/src/main/java/ar/com/reduceFatFast/model/DietaSemanal.java
@@ -26,27 +26,27 @@ public @Data class DietaSemanal {
 	public boolean validacion;
 	
 	@ManyToMany(cascade=CascadeType.PERSIST) @JoinColumn(name="id") @MapKey(name="id")
-	private Map<Integer,Dia> dias;
+	private Map<Long,Dia> dias;
 	
 	public DietaSemanal () {
-		this.setDias(new HashMap<Integer,Dia>());
+		this.setDias(new HashMap<Long,Dia>());
 	}
 	
-	public void agregarComida (Comida unaComida, Integer numeroDia, ComidaDelDia comidaDelDia){
+	public void agregarComida (Comida unaComida, Long numeroDia, ComidaDelDia comidaDelDia){
 		if (dias.containsKey(numeroDia) != true )
 			dias.put(numeroDia,new Dia());
 		dias.get(numeroDia).setComida(comidaDelDia, unaComida);
 	}
 	
-	public void agregarDia(Dia dia, Integer pos){
+	public void agregarDia(Dia dia, Long pos){
 		this.getDias().put(pos, dia);
 	}
 	
-	public void agregarDias(Map<Integer,Dia> dias){
+	public void agregarDias(Map<Long,Dia> dias){
 		this.getDias().putAll(dias);
 	}
 
-	public Dia getDia(Integer j) {
+	public Dia getDia(Long j) {
 		return this.getDias().get(j);
 	}
